Use redux for the export of csv/json
