+++ src/constants/ExportConstants.js
@@ -1,4 +1,4 @@
 export const FORMING_CSV = 'FORMING_CSV';
 export const FORMING_JSON = 'FORMING_JSON';
 export const SET_CSV_ERROR_MESSAGE = 'SET_CSV_ERROR_MESSAGE';
-export const SET_JSON_ERROR_MESSAG = 'SET_JSON_ERROR_MESSAG';
+export const SET_JSON_ERROR_MESSAGE = 'SET_JSON_ERROR_MESSAGE';
